Importer metrics

Importer metrics

cp_importer(
  dataset_key = NULL,
  state = NULL,
  running = FALSE,
  start = 0,
  limit = 10,
  ...
)

Arguments

dataset_key
(character) a dataset key to filter by. optional

state
(character) filter listed import metrics by their state, e.g. the last failed import. one of: downloading, processing, inserting, unchanged, finished, canceled, failed. optional

running
(logical) if only a list of running imports should be returned. default: FALSE. optional

start
(integer) requested number of offset records. Default: 0

limit
(integer) requested number of maximum records to be returned. Default: 10; max: 1000

...
curl options passed on to verb-GET

Value

a named list, with slots offset (integer), limit (integer), total (integer), result (list), empty (boolean), and last (boolean). The result slot is a list itself, with any number of results as named lists

Examples

if (cp_up("/importer?limit=1")) {
cp_importer(limit = 1)
}

#> $offset
#> [1] 0
#> 
#> $limit
#> [1] 1
#> 
#> $total
#> [1] 2812
#> 
#> $result
#> $result[[1]]
#> $result[[1]]$datasetKey
#> [1] 1093
#> 
#> $result[[1]]$attempt
#> [1] 13
#> 
#> $result[[1]]$job
#> [1] "ImportJob"
#> 
#> $result[[1]]$state
#> [1] "finished"
#> 
#> $result[[1]]$started
#> [1] "2021-07-15T19:49:57.677555"
#> 
#> $result[[1]]$finished
#> [1] "2021-07-15T19:50:00.209621"
#> 
#> $result[[1]]$createdBy
#> [1] 103
#> 
#> $result[[1]]$bareNameCount
#> [1] 0
#> 
#> $result[[1]]$distributionCount
#> [1] 40
#> 
#> $result[[1]]$estimateCount
#> [1] 0
#> 
#> $result[[1]]$mediaCount
#> [1] 0
#> 
#> $result[[1]]$nameCount
#> [1] 68
#> 
#> $result[[1]]$referenceCount
#> [1] 31
#> 
#> $result[[1]]$synonymCount
#> [1] 9
#> 
#> $result[[1]]$taxonCount
#> [1] 59
#> 
#> $result[[1]]$treatmentCount
#> [1] 0
#> 
#> $result[[1]]$typeMaterialCount
#> [1] 6
#> 
#> $result[[1]]$vernacularCount
#> [1] 0
#> 
#> $result[[1]]$distributionsByGazetteerCount
#> $result[[1]]$distributionsByGazetteerCount$mrgid
#> [1] 40
#> 
#> 
#> $result[[1]]$issuesCount
#> $result[[1]]$issuesCount$`not interpreted`
#> [1] 42
#> 
#> $result[[1]]$issuesCount$`self referenced relation`
#> [1] 42
#> 
#> $result[[1]]$issuesCount$`citation details unparsed`
#> [1] 25
#> 
#> $result[[1]]$issuesCount$`escaped characters`
#> [1] 14
#> 
#> $result[[1]]$issuesCount$`published before genus`
#> [1] 8
#> 
#> $result[[1]]$issuesCount$`parent id invalid`
#> [1] 4
#> 
#> $result[[1]]$issuesCount$`citation authors unparsed`
#> [1] 3
#> 
#> $result[[1]]$issuesCount$`reference id invalid`
#> [1] 2
#> 
#> $result[[1]]$issuesCount$`unmatched reference brackets`
#> [1] 2
#> 
#> 
#> $result[[1]]$namesByCodeCount
#> $result[[1]]$namesByCodeCount$zoological
#> [1] 62
#> 
#> 
#> $result[[1]]$namesByRankCount
#> $result[[1]]$namesByRankCount$species
#> [1] 51
#> 
#> $result[[1]]$namesByRankCount$genus
#> [1] 7
#> 
#> $result[[1]]$namesByRankCount$family
#> [1] 4
#> 
#> $result[[1]]$namesByRankCount$kingdom
#> [1] 1
#> 
#> $result[[1]]$namesByRankCount$phylum
#> [1] 1
#> 
#> $result[[1]]$namesByRankCount$subphylum
#> [1] 1
#> 
#> $result[[1]]$namesByRankCount$class
#> [1] 1
#> 
#> $result[[1]]$namesByRankCount$subclass
#> [1] 1
#> 
#> $result[[1]]$namesByRankCount$order
#> [1] 1
#> 
#> 
#> $result[[1]]$namesByStatusCount
#> $result[[1]]$namesByStatusCount$established
#> [1] 53
#> 
#> 
#> $result[[1]]$namesByTypeCount
#> $result[[1]]$namesByTypeCount$scientific
#> [1] 68
#> 
#> 
#> $result[[1]]$nameRelationsByTypeCount
#> $result[[1]]$nameRelationsByTypeCount$basionym
#> [1] 9
#> 
#> $result[[1]]$nameRelationsByTypeCount$type
#> [1] 7
#> 
#> 
#> $result[[1]]$synonymsByRankCount
#> $result[[1]]$synonymsByRankCount$species
#> [1] 9
#> 
#> 
#> $result[[1]]$taxaByRankCount
#> $result[[1]]$taxaByRankCount$species
#> [1] 42
#> 
#> $result[[1]]$taxaByRankCount$genus
#> [1] 7
#> 
#> $result[[1]]$taxaByRankCount$family
#> [1] 4
#> 
#> $result[[1]]$taxaByRankCount$kingdom
#> [1] 1
#> 
#> $result[[1]]$taxaByRankCount$phylum
#> [1] 1
#> 
#> $result[[1]]$taxaByRankCount$subphylum
#> [1] 1
#> 
#> $result[[1]]$taxaByRankCount$class
#> [1] 1
#> 
#> $result[[1]]$taxaByRankCount$subclass
#> [1] 1
#> 
#> $result[[1]]$taxaByRankCount$order
#> [1] 1
#> 
#> 
#> $result[[1]]$typeMaterialByStatusCount
#> $result[[1]]$typeMaterialByStatusCount$holotype
#> [1] 6
#> 
#> 
#> $result[[1]]$usagesByStatusCount
#> $result[[1]]$usagesByStatusCount$accepted
#> [1] 59
#> 
#> $result[[1]]$usagesByStatusCount$synonym
#> [1] 9
#> 
#> 
#> $result[[1]]$usagesByOriginCount
#> $result[[1]]$usagesByOriginCount$source
#> [1] 62
#> 
#> $result[[1]]$usagesByOriginCount$`denormed classification`
#> [1] 6
#> 
#> 
#> $result[[1]]$downloadUri
#> [1] "http://www.marinespecies.org/export/coldp/WoRMS_Thermosbaenacea.zip"
#> 
#> $result[[1]]$origin
#> [1] "external"
#> 
#> $result[[1]]$download
#> [1] "2021-07-01T01:34:57"
#> 
#> $result[[1]]$md5
#> [1] "ABC32EF82E266D565A54AC6125906FC7"
#> 
#> $result[[1]]$verbatimCount
#> [1] 256
#> 
#> $result[[1]]$verbatimByTermCount
#> $result[[1]]$verbatimByTermCount$`col:Taxon`
#> [1] 53
#> 
#> $result[[1]]$verbatimByTermCount$`col:TypeMaterial`
#> [1] 6
#> 
#> $result[[1]]$verbatimByTermCount$`col:Distribution`
#> [1] 37
#> 
#> $result[[1]]$verbatimByTermCount$`col:Synonym`
#> [1] 9
#> 
#> $result[[1]]$verbatimByTermCount$`col:NameRelation`
#> [1] 58
#> 
#> $result[[1]]$verbatimByTermCount$`col:Name`
#> [1] 62
#> 
#> $result[[1]]$verbatimByTermCount$`col:Reference`
#> [1] 31
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:ID`
#> [1] 62
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:code`
#> [1] 62
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:link`
#> [1] 62
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:rank`
#> [1] 62
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:genus`
#> [1] 51
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:status`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:remarks`
#> [1] 9
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:uninomial`
#> [1] 11
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:authorship`
#> [1] 61
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:basionymID`
#> [1] 51
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:referenceID`
#> [1] 61
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:scientificName`
#> [1] 62
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:publishedInPage`
#> [1] 4
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:publishedInYear`
#> [1] 61
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Name`$`col:specificEpithet`
#> [1] 51
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:ID`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:link`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:class`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:genus`
#> [1] 49
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:order`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:family`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:nameID`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:phylum`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:kingdom`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:species`
#> [1] 42
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:parentID`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:subclass`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:subphylum`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:environment`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:provisional`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:referenceID`
#> [1] 52
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:scrutinizer`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:scrutinizerID`
#> [1] 53
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Taxon`$`col:scrutinizerDate`
#> [1] 53
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Synonym`
#> $result[[1]]$verbatimByRowTypeCount$`col:Synonym`$`col:link`
#> [1] 9
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Synonym`$`col:nameID`
#> [1] 9
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Synonym`$`col:status`
#> [1] 9
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Synonym`$`col:taxonID`
#> [1] 9
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:ID`
#> [1] 31
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:doi`
#> [1] 6
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:link`
#> [1] 31
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:year`
#> [1] 31
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:title`
#> [1] 31
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:author`
#> [1] 31
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:source`
#> [1] 24
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:details`
#> [1] 25
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:remarks`
#> [1] 4
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Reference`$`col:citation`
#> [1] 31
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:area`
#> [1] 37
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:areaID`
#> [1] 37
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:status`
#> [1] 1
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:remarks`
#> [1] 1
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:taxonID`
#> [1] 37
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:gazetteer`
#> [1] 37
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:Distribution`$`col:referenceID`
#> [1] 2
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:NameRelation`
#> $result[[1]]$verbatimByRowTypeCount$`col:NameRelation`$`col:type`
#> [1] 58
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:NameRelation`$`col:nameID`
#> [1] 58
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:NameRelation`$`col:referenceID`
#> [1] 50
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:NameRelation`$`col:relatedNameID`
#> [1] 58
#> 
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:ID`
#> [1] 6
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:link`
#> [1] 6
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:nameID`
#> [1] 6
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:status`
#> [1] 6
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:latitude`
#> [1] 5
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:locality`
#> [1] 5
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:longitude`
#> [1] 5
#> 
#> $result[[1]]$verbatimByRowTypeCount$`col:TypeMaterial`$`col:referenceID`
#> [1] 3
#> 
#> 
#> 
#> $result[[1]]$upload
#> [1] FALSE
#> 
#> $result[[1]]$nameRelationsCount
#> [1] 16
#> 
#> $result[[1]]$usagesCount
#> [1] 68
#> 
#> $result[[1]]$taxonConceptRelationsCount
#> [1] 0
#> 
#> $result[[1]]$speciesInteractionsCount
#> [1] 0
#> 
#> 
#> 
#> $empty
#> [1] FALSE
#> 
#> $last
#> [1] FALSE
#>