rcol: R Client for the Catalogue of Life / ChecklistBank API

Access the Catalogue of Life (COL) through the ChecklistBank (CLB) web API https://api.checklistbank.org. Match scientific names against any public dataset or the latest COL release, look up datasets, taxa and name usages, navigate the taxonomic tree, run the ChecklistBank name and value parsers, and discover the latest annual and monthly COL base and extended releases. Modelled on the conventions of 'rgbif' and 'taxadb'.

See also

Author

Maintainer: Markus Döring mdoering@gbif.org (ORCID)

Authors:

Markus Döring mdoering@gbif.org (ORCID)

Paul Efren Santos Andrade paulefrens@gmail.com (ORCID)

Other contributors:

Catalogue of Life [copyright holder, funder]